requirements.txt
requests
langchain
black
openai
install
google-search-results
tweepy
flask
langchain-openai
python-dotenv
langchain-community
langchain-pinecone
langchainhub
pydantic
langchain-core
pylint
beautifulsoup4
tiktoken
pinecone-client
unstructured
nltk
python-magic-bin
fastapi
jinja2
uvicorn
streamlit
streamlit-chat
tqdm
pypdf

